test(core): convert test core to typescript and fix code
